pyproject.toml
[project]
name = "pytorch-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "scipy>=1.16.2",
    "setuptools>=80.9.0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff.lint]
select = ["D", "I", "W", "E", "F"]
ignore = ["D100", "D104", "D105", "D107"]
